requirements.txt
torch==2.2.2
torchio>=0.19.5
hydra-core==1.3.2
omegaconf==2.3.0
tensorboard==2.16.2
tqdm==4.66.4
torchvision==0.17.2
numpy==1.26.4
monai>=1.3.0
torchinfo==1.8.0
einops>=0.6.0
nibabel>=5.0.0
pytorch_msssim>=0.2.1
tensorboard>=2.16.2
jupyter>=1.0.0
lpips>=0.1.4
umap-learn>=0.5.0